refactor(catalog-cli): carve links command family into catalog_cmds (nexus-kgyoz)
Third deferred kgyoz CLI carve, the largest. Lift the link command family
out of commands/catalog.py into commands/catalog_cmds/links.py: link,
unlink, links, link-bulk-delete, link-audit, links-for-file, link-density,
suggest-links, generate-links, link-generate (10 commands), plus the two
link-only render helpers _endpoint_label / _unique_edges_by_target that only
'links' uses. Net ~620 lines + 2 helpers out of the god module.

Mirrors the seam-3/owners + backfill pattern: plain @click.command +
register(group); commands.catalog calls _links_cmds.register(catalog).
'nx catalog link' (etc.) resolve identically — names, options, output
unchanged; 47 total catalog commands preserved.

Behaviour-preserving: bodies moved verbatim (incl. lazy imports for
snapshot_links / link_generator / statistics / Path, and link-generate's
ctx.invoke of generate-links). Shared helpers _get_catalog /
_get_catalog_writer / _resolve_tumbler reached via the nexus.commands.catalog
module object inside each body — import stays acyclic (no module-level
import of commands.catalog) and the _get_catalog patch seam holds. The one
test importing _endpoint_label by symbol updated to the new location.

Four contract pins in test_catalog_cli.py::TestKgyozLinksCarve: all-10
registered, callbacks-in-carved-module (__module__ check, anti-re-inline),
link-only-helpers-moved, and a behavioural module-routed _get_catalog pin
via link-audit. Guard suites green (191 across cli/links/xnz0o/doctor).
